# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service 'Short blog API'",
        "version": "0.1"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List of users",
                "description": "List of users",
                "operationId": "/api/users(GET)",
                "parameters": [
                    {
                        "name": "sort_column",
                        "in": "query",
                        "description": "Sort column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Users per page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keys",
                        "in": "query",
                        "description": "User keys",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "",
                            "example": ""
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name and email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": ""
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Includes",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "posts"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list of users",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "operationId": "/api/users(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return model of new user",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id",
                "description": "Show the specified resource from storage.",
                "operationId": "/api/users/{id}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": ""
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Includes",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "posts"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get exists post"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "/api/users/{id}(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return post model",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Remove the specified resource from storage.",
                "operationId": "/api/users/{id}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete exists post",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            }
        },
        "/api/user-change-status/{id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Change user status",
                "description": "Change status",
                "operationId": "/api/user-change-status/{id}(PATCH)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": ""
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserChangeStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Change post status",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserChangeStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "description": "Post status",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": ""
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": ""
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": ""
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": ""
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "0"
                    }
                },
                "type": "object"
            }
        }
    }
}